// frontend/src/lib/api/models/channelAdminQuickCreateRequest.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Stream Archive API
 * 멀티 플랫폼 스트리밍 녹화 시스템 API
 * OpenAPI spec version: 1.0.0
 */
import type { ChannelAdminQuickCreateRequestContentPrivacy } from "./channelAdminQuickCreateRequestContentPrivacy";
import type { ChannelAdminQuickCreateRequestPlatformType } from "./channelAdminQuickCreateRequestPlatformType";
import type { ScheduleRequest } from "./scheduleRequest";

/**
 * 간편 채널 생성 요청 (관리자)
 */
export interface ChannelAdminQuickCreateRequest {
  /** 채널 이름 */
  name: string;
  /** 콘텐츠 공개 범위 (PUBLIC/UNLISTED/PRIVATE) */
  contentPrivacy: ChannelAdminQuickCreateRequestContentPrivacy;
  /** 플랫폼 유형 (CHZZK/TWITCH/SOOP/YOUTUBE) */
  platformType: ChannelAdminQuickCreateRequestPlatformType;
  /** 플랫폼 채널 ID */
  platformChannelId: string;
  /** 프로필 동기화 여부 */
  isSyncProfile: boolean;
  schedule?: ScheduleRequest;
}