// frontend/src/lib/api/models/channelAdminQuickCreateRequestContentPrivacy.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Stream Archive API
 * 멀티 플랫폼 스트리밍 녹화 시스템 API
 * OpenAPI spec version: 1.0.0
 */

/**
 * 콘텐츠 공개 범위 (PUBLIC/UNLISTED/PRIVATE)
 */
export type ChannelAdminQuickCreateRequestContentPrivacy =
  (typeof ChannelAdminQuickCreateRequestContentPrivacy)[keyof typeof ChannelAdminQuickCreateRequestContentPrivacy];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ChannelAdminQuickCreateRequestContentPrivacy = {
  PUBLIC: "PUBLIC",
  UNLISTED: "UNLISTED",
  PRIVATE: "PRIVATE",
} as const;