// frontend/src/lib/api/models/scheduleRequestScheduleType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Stream Archive API
 * 멀티 플랫폼 스트리밍 녹화 시스템 API
 * OpenAPI spec version: 1.0.0
 */

/**
 * 녹화 스케줄 유형 (ONCE/ALWAYS/N_DAYS_OF_EVERY_WEEK/SPECIFIC_DAY)
 */
export type ScheduleRequestScheduleType =
  (typeof ScheduleRequestScheduleType)[keyof typeof ScheduleRequestScheduleType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ScheduleRequestScheduleType = {
  ONCE: "ONCE",
  ALWAYS: "ALWAYS",
  N_DAYS_OF_EVERY_WEEK: "N_DAYS_OF_EVERY_WEEK",
  SPECIFIC_DAY: "SPECIFIC_DAY",
} as const;