// frontend/src/lib/api/models/scheduleRequestRecordQuality.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Stream Archive API
 * 멀티 플랫폼 스트리밍 녹화 시스템 API
 * OpenAPI spec version: 1.0.0
 */

/**
 * 녹화 화질
 */
export type ScheduleRequestRecordQuality =
  (typeof ScheduleRequestRecordQuality)[keyof typeof ScheduleRequestRecordQuality];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ScheduleRequestRecordQuality = {
  BEST: "BEST",
  P2160_60: "P2160_60",
  P2160: "P2160",
  P1440_60: "P1440_60",
  P1440: "P1440",
  P1080_60: "P1080_60",
  P1080: "P1080",
  P720_60: "P720_60",
  P720: "P720",
  P480: "P480",
  P240: "P240",
  P144: "P144",
  WORST: "WORST",
} as const;